Added Stability Map

+++ models/llm_interface.py
@@ -16,7 +16,7 @@ class LLMInterface:
 
         self.model = AutoModelForCausalLM.from_pretrained(
             model_name,
-            torch_dtype=torch.float16 if self.device == "cuda" else torch.float32,
+            torch_dtype=torch.float16 if self.device == "cuda" else torch.float16,
             device_map="auto"
         )
